Send IntelliJ productCode during agent initialization
