+++ src/main/kotlin/com/sourcegraph/cody/telemetry/TelemetryV2.kt
@@ -6,39 +6,20 @@ import com.sourcegraph.cody.agent.CodyAgentService
 import com.sourcegraph.cody.agent.protocol.BillingMetadata
 import com.sourcegraph.cody.agent.protocol.TelemetryEvent
 import com.sourcegraph.cody.agent.protocol.TelemetryEventParameters
+import com.sourcegraph.config.ConfigUtil
 
 class TelemetryV2 {
   companion object {
-    private val intellijProductCodeMap =
-        mapOf(
-            "IU" to 1L, // IntelliJ IDEA Ultimate
-            "IC" to 2L, // IntelliJ IDEA Community
-            "IE" to 3L, // IntelliJ IDEA Educational
-            "PS" to 4L, // PhpStorm
-            "WS" to 5L, // WebStorm
-            "PY" to 6L, // PyCharm Professional
-            "PC" to 7L, // PyCharm Community
-            "PE" to 8L, // PyCharm Educational
-            "RM" to 9L, // RubyMine
-            "OC" to 10L, // AppCode
-            "CL" to 11L, // CLion
-            "GO" to 12L, // GoLand
-            "DB" to 13L, // DataGrip
-            "RD" to 14L, // Rider
-            "AI" to 15L, // Android Studio
-        )
-
     fun sendTelemetryEvent(
         project: Project,
         feature: String,
         action: String,
         parameters: TelemetryEventParameters? = null
     ) {
-      val build = ApplicationInfo.getInstance().build
       val versionParameters =
           mapOf(
-              "ideProductCode" to intellijProductCodeMap.getOrDefault(build.productCode, 0L),
-              "ideBaselineVersion" to build.baselineVersion.toLong())
+              "ideProductCode" to ConfigUtil.getIntellijProductCode(),
+              "ideBaselineVersion" to ApplicationInfo.getInstance().build.baselineVersion.toLong())
       val baseParameters = parameters ?: TelemetryEventParameters()
       val newParameters =
           baseParameters.copy(metadata = baseParameters.metadata?.plus(versionParameters))
